#修改了许多NumericUpDown自输入小数时，其表面显示的数字(int)和实际的value值不一样的问题(全局、摇麦、主页的各种td值）。

+++ LightController/Common/IniFileAst.cs
@@ -116,7 +116,7 @@ namespace LightController.Common
 			}
 			catch (Exception) {
 				double tempValue = Double.Parse(Ival.ToString());
-				value = Convert.ToInt32(tempValue);
+				value = Convert.ToInt16(tempValue);
 			}
 
 			WritePrivateProfileString(section, name, value.ToString(), this.filePath);
